nrf52840/Packetbuilder/requirements.txt
parse
scapy
pyserial
pycryptodome